fixed some performance

+++ test/Test.Alea.CUDA.Extension.MGPU/Program.fs
@@ -1,1 +1,1 @@
-﻿Test.Alea.CUDA.Extension.MGPU.BulkRemove.``BulkRemove moderngpu benchmark : float``()
+﻿Test.Alea.CUDA.Extension.MGPU.BulkRemove.``BulkRemove moderngpu benchmark : int``()
